requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4 